Вынес часть кода из OfferScreen в отдельный компонент Offer
