+++ src/components/app/app.tsx
@@ -1,5 +1,4 @@
 import MainScreen from '../../pages/main-screen/main-screen';
-//import OfferScreen from '../../pages/offer-screen/offer-screen';
 
 type AppProps = {
   placesToStayCount: number;
@@ -8,7 +7,6 @@ type AppProps = {
 function App({placesToStayCount}: AppProps): JSX.Element {
   return (
     <MainScreen placesToStayCount={placesToStayCount}/>
-    //<OfferScreen/>
   );
 }
